+++ src/lib/typings.ts
@@ -1,7 +1,13 @@
-import { FetchAction } from '../state/actions';
+import { AnyAction, Action } from 'redux';
+import { ThunkAction } from 'redux-thunk';
 import { ActionTypes, PrHubState } from '../state/types';
 
 export type Enum<T extends object> = T[keyof T];
 
-type ReducerAction = [Enum<typeof ActionTypes>, () => PrHubState];
+export type ReducerAction = [Enum<typeof ActionTypes>, () => PrHubState];
 export type SplitReducer = (state: PrHubState, action: FetchAction) => ReducerAction[];
+
+export type Task<Args = undefined, ReturnType = void> = (
+  ...args: Args[]
+) => ThunkAction<ReturnType, PrHubState, null, AnyAction>;
+export type FetchAction = Action & { payload?: any };